+++ src/standard/webgl/webgl_system.ts
@@ -28,6 +28,8 @@ import Component from "../../component/component";
 import Renderable from "../../rendering/renderable";
 import IScene from "../../scene/iscene";
 import SystemEntity from "../../system/system_entity";
+import ImageSystem from "../image/image_system";
+import ImageAsset from "../image/image_asset";
 
 class WebGLSystem extends System {
 
@@ -35,24 +37,39 @@ class WebGLSystem extends System {
 
     private static readonly VERTEX_SHADER = `#version 300 es
         in vec2 aVertexPosition;
+        in vec2 aTexturePosition;
 
         uniform mat4 uViewMatrix;
         uniform mat4 uModelMatrix;
         uniform mat4 uProjectionMatrix;
 
+        out vec2 vTextureCoordinate;
+
         void main() {
             gl_Position = uProjectionMatrix * uViewMatrix * uModelMatrix * vec4(aVertexPosition, 0, 1);
+            vTextureCoordinate = aTexturePosition;
         }
     `;
     private static readonly FRAGMENT_SHADER = `#version 300 es
         precision mediump float;
 
+        uniform bool uTexturePresent;
         uniform vec4 uColor;
+        uniform sampler2D uTexture;
+
+        in vec2 vTextureCoordinate;
 
         out vec4 outColor;
 
         void main() {
-            outColor = uColor;
+            vec4 texturedColor;
+            if (uTexturePresent) {
+                texturedColor = texture(uTexture, vTextureCoordinate);
+            }
+            else {
+                texturedColor = uColor;
+            }
+            outColor = texturedColor;
         }
     `;
 
@@ -65,6 +82,8 @@ class WebGLSystem extends System {
     private gl: WebGL2RenderingContext;
     private program: WebGLProgram | null;
     private renderables: Renderable[];
+    private placeholderTexture: WebGLTexture;
+    private textures: Record<string, WebGLTexture>;
 
     constructor(messageBus: IMessageBus, gl: WebGL2RenderingContext, { scene, entities, subscriberID, renderables }:
         { scene: IScene | undefined; entities: SystemEntity[]; subscriberID: number | undefined; renderables: Renderable[] } =
@@ -72,10 +91,18 @@ class WebGLSystem extends System {
         super(messageBus, { evaluator: WebGLSystem.EVALUATOR, scene, entities, subscriberID });
         this.gl = gl;
         this.renderables = renderables;
-        this.messageBus.Subscribe(this, [Game.MESSAGE_RENDER, WebGLSystem.MESSAGE_LOAD_RENDERABLES]);
+        this.messageBus.Subscribe(this, [Game.MESSAGE_RENDER, WebGLSystem.MESSAGE_LOAD_RENDERABLES, ImageSystem.MESSAGE_FINISH_LOAD]);
         const vertexShader = this.createShader(gl.VERTEX_SHADER, WebGLSystem.VERTEX_SHADER);
 		const fragmentShader = this.createShader(gl.FRAGMENT_SHADER, WebGLSystem.FRAGMENT_SHADER);
-		this.program = this.createProgram(vertexShader, fragmentShader);
+        this.program = this.createProgram(vertexShader, fragmentShader);
+        const placeholderTexture = gl.createTexture();
+        gl.bindTexture(gl.TEXTURE_2D, placeholderTexture);
+        gl.texImage2D(gl.TEXTURE_2D, 0, gl.RGBA, 1, 1, 0, gl.RGBA, gl.UNSIGNED_BYTE, new Uint8Array([0,0,255,255]));
+        if (placeholderTexture === null) {
+            throw("Failed to create placeholder texture");
+        }
+        this.placeholderTexture = placeholderTexture;
+        this.textures = {};
     }
 
     public OnMessage(message: IMessage): void {
@@ -97,10 +124,37 @@ class WebGLSystem extends System {
                 this.renderables.push(...renderMessage.payload);
                 break;
             }
+            case ImageSystem.MESSAGE_FINISH_LOAD: {
+                const loadMessage = message as Message<ImageAsset>;
+                if (loadMessage.payload == undefined) {
+                    return;
+                }
+                if (!loadMessage.payload.success) {
+                    return;
+                }
+                const gl = this.gl;
+                const level = 0;
+                const internalFormat = gl.RGBA;
+                const srcFormat = gl.RGBA;
+                const srcType = gl.UNSIGNED_BYTE;
+                const glTexture = gl.createTexture();
+        
+                gl.bindTexture(gl.TEXTURE_2D, glTexture);
+                gl.texImage2D(gl.TEXTURE_2D, level, internalFormat, srcFormat, srcType, loadMessage.payload.image);
+                gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_WRAP_S, gl.CLAMP_TO_EDGE);
+                gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_WRAP_T, gl.CLAMP_TO_EDGE);
+                gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_MIN_FILTER, gl.LINEAR);
+                gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_MAG_FILTER, gl.LINEAR);
+                gl.generateMipmap(gl.TEXTURE_2D);
+                if (glTexture === null) {
+                    throw("Failed to create texture");
+                }
+                this.textures[loadMessage.payload.name] = glTexture;
+            }
         }
     }
 
-    createShader(type: number, source: string): WebGLShader {
+    private createShader(type: number, source: string): WebGLShader {
         const gl = this.gl;
         const shader = gl.createShader(type);
         if (!shader) {
@@ -116,7 +170,7 @@ class WebGLSystem extends System {
 		return shader;
 	}
 
-	createProgram(vertex: WebGLShader, fragment: WebGLShader): WebGLProgram {
+	private createProgram(vertex: WebGLShader, fragment: WebGLShader): WebGLProgram {
         const gl = this.gl;
         const program = gl.createProgram();
         if (!program) {
@@ -181,7 +235,6 @@ class WebGLSystem extends System {
             );
             
             gl.enable(gl.SCISSOR_TEST);
-            gl.enable(gl.BLEND);
     
             // Clear the screen
             gl.clearDepth(1.0);
@@ -201,7 +254,7 @@ class WebGLSystem extends System {
     
             // Enable alpha blending
             gl.enable(gl.BLEND);
-            gl.blendFunc(gl.ONE, gl.ONE_MINUS_SRC_ALPHA);
+            gl.blendFunc(gl.SRC_ALPHA, gl.ONE_MINUS_SRC_ALPHA);
 
             const viewMatrix = new Matrix4D();
             viewMatrix.Translate(transform.position.Invert());
@@ -214,6 +267,9 @@ class WebGLSystem extends System {
             const viewLocation = gl.getUniformLocation(program, "uViewMatrix");
             const modelLocation = gl.getUniformLocation(program, "uModelMatrix");
             const projectionLocation = gl.getUniformLocation(program, "uProjectionMatrix");
+            const texturePresentLocation = gl.getUniformLocation(program, "uTexturePresent");
+            const textureLocation = gl.getUniformLocation(program, "uTexture");
+
 
 			gl.uniformMatrix4fv(
 				viewLocation,
@@ -230,19 +286,45 @@ class WebGLSystem extends System {
             
             for (const renderable of this.renderables) {
                 const vao = gl.createVertexArray();
-                const position = gl.getAttribLocation(program, "aVertexPosition");
+                const positionLocation = gl.getAttribLocation(program, "aVertexPosition");
+                const texturePositionLocation = gl.getAttribLocation(program, "aTexturePosition");
                 const positionBuffer = gl.createBuffer();
+                const textureBuffer = gl.createBuffer();
+                let texturePresent = false;
+                let texturePoints = renderable.verticies;
+                let texture = this.placeholderTexture;
+                if (renderable.texture !== undefined) {
+                    texturePresent = true;
+                    texturePoints = renderable.texture.points;
+                    texture = this.textures[renderable.texture.image];
+                    if (texture === undefined) {
+                        texture = this.placeholderTexture;
+                    }
+                }
+
+                gl.uniform1i(texturePresentLocation, Number(texturePresent))
+				gl.activeTexture(gl.TEXTURE0);
+                gl.bindTexture(gl.TEXTURE_2D, texture);
+				gl.uniform1i(textureLocation, 0);
         
                 // bind vao
                 gl.bindVertexArray(vao);
                 // Enable attribute
-                gl.enableVertexAttribArray(position);
+                gl.enableVertexAttribArray(positionLocation);
                 // bind buffer
                 gl.bindBuffer(gl.ARRAY_BUFFER, positionBuffer);
-                gl.vertexAttribPointer(position, 2, gl.FLOAT, false, 0, 0);
+                gl.vertexAttribPointer(positionLocation, 2, gl.FLOAT, false, 0, 0);
         
                 // send buffer data
                 gl.bufferData(gl.ARRAY_BUFFER, renderable.verticies, gl.DYNAMIC_DRAW);
+
+                gl.bindBuffer(gl.ARRAY_BUFFER, textureBuffer);
+
+                gl.bufferData(gl.ARRAY_BUFFER, texturePoints, gl.DYNAMIC_DRAW);
+
+                gl.enableVertexAttribArray(texturePositionLocation);
+
+                gl.vertexAttribPointer(texturePositionLocation, 2, gl.FLOAT, false, 0, 0);
 
                 gl.uniformMatrix4fv(
                     modelLocation,